fixed the deletes

+++ server/controllers/BlogController.js
@@ -64,8 +64,7 @@ export default class BlogController {
 
   async edit(req, res, next) {
     try {
-      let data = await _blogService.findOneAndUpdate({ _id: req.params.id, },
-        req.body, { new: true })
+      let data = await _blogService.findOneAndUpdate({ _id: req.params.id, authorId: req.session.uid }, req.body, { new: true })
       if (data) {
         return res.send(data)
       }
@@ -77,7 +76,10 @@ export default class BlogController {
 
   async delete(req, res, next) {
     try {
-      await _blogService.findOneAndRemove({ _id: req.params.id })
+      let data = await _blogService.findOneAndRemove({ _id: req.params.id, authorId: req.session.uid })
+      if (!data) {
+        throw new Error("invalid id")
+      }
       res.send("Deleted Blog")
     } catch (error) {
       next(error)
